Refactored displaying as a table into new file.
